refactored usePlaces to repository Closes #24
