+++ app/composables/usePlaces.ts
@@ -1,7 +1,4 @@
-import maplibregl from "maplibre-gl";
 import type { Place } from "~~/shared/place.type";
-
-let isInitialLoad = true;
 
 export function usePlaces() {
   const api = useApi();
@@ -9,30 +6,6 @@ export function usePlaces() {
   const query = useQuery();
 
   const places = useState<Place[]>("places", () => []);
-  const lastQueryLocation = useState<{
-    latitude: number;
-    longitude: number;
-  } | null>("places:lastLocation", () => null);
-
-  function shouldQueryNewPlaces(newLocation: {
-    latitude: number;
-    longitude: number;
-  }) {
-    if (!lastQueryLocation.value) return true;
-
-    const point1 = new maplibregl.LngLat(
-      newLocation.longitude,
-      newLocation.latitude,
-    );
-    const point2 = new maplibregl.LngLat(
-      lastQueryLocation.value.longitude,
-      lastQueryLocation.value.latitude,
-    );
-
-    const distanceInMeters = point1.distanceTo(point2);
-
-    return distanceInMeters > 500;
-  }
 
   async function fetchPlaces(location: {
     latitude: number;
@@ -41,7 +14,6 @@ export function usePlaces() {
     const data = await api.fetchPlaces(location);
 
     places.value = data;
-    lastQueryLocation.value = location;
   }
 
   async function searchPlaces(
@@ -54,35 +26,24 @@ export function usePlaces() {
     const data = await api.searchPlaces(location, query);
 
     places.value = data;
-    lastQueryLocation.value = location;
   }
 
   const throttledFetchPlaces = useThrottleFn(fetchPlaces, 1000, true);
   const throttledSearchPlaces = useThrottleFn(searchPlaces, 500, true);
 
-  if (isInitialLoad) {
-    isInitialLoad = false;
-    watch(
-      () => ({
-        latitude: userLocation.value.latitude,
-        longitude: userLocation.value.longitude,
-        query: query.value,
-      }),
-      async (newLocation, oldLocation) => {
-        if (query.value.length < 2) {
-          if (
-            shouldQueryNewPlaces(newLocation) ||
-            oldLocation?.query != newLocation.query
-          ) {
-            await throttledFetchPlaces(newLocation);
-          }
-        } else {
-          await throttledSearchPlaces(newLocation, query.value);
-        }
-      },
-      { immediate: true },
-    );
+  async function fetch(): Promise<void> {
+    if (query.value
+    ) {
+      await throttledSearchPlaces(userLocation.value, query.value);
+
+    }
+    else {
+      await throttledFetchPlaces(userLocation.value);
+    }
   }
 
-  return places;
+  return {
+    places: readonly(places),
+    fetch
+  };
 }
